Refactor account setup to "ensures"
That is, instead of setting display name and saving session at register time, we do it as an ensure-y operation on every login.

This will help us recover from partial setups and gives us a pattern for how to upgrade accounts with new logic instead of depending on it happening at registration time.