Load stream state once access token is ready
We were running into a race condition where, if you had a saved session, we would set it up before loading stream state - but if it was a new session, stream state would try to make the request before your access token was ready.

Now, instead of having the caller ask for this, we have the Playroom do it internally! (We'll probably move more and more stuff in there tbh…)