Big refactor stuff into the model layer
The Playroom class is now more responsible for the chat and stream state management, and the view is just a bridge to show it.

This led us to create a couple more layers inside the model thing too, to split up app state from our matrix logic from the matrix network calls.