Move roomId into a settings.js file
Feels like this is gonna be way easier to manage!

I initially had it in a settings.json to be a bit of a simpler file type, but then I realized that comments were gonna be a big deal, so .js is actually just by far the easiest way to do it.