Script to build hydrogen-web
I want to try to move off build tools like Vite altogether? This is an important step to it.

I noticed when running `esbuild` with `--analyze` that two test libraries were huge parts of the bundle. I excluded them, and cut the bundle size in half! Still higher than I want at nearly 800K, but it's still good news! (Also not the worst for Hydrogen to be bigger, so it can work while we do the other stuff. Oh right, it should be an async import maybe!)