Fix security check for broadcast events
Okay, right, we're not using widgets anymore, so the event type needs to change.

But I also learned about the `users_default` and `state_default` settings, which control the settings when the broadcast event isn't specified! This will make it easier for people to edit these settings in their own clients, and means that it comes with pretty sensible defaults, which is nice!